Fix three frail tests

+++ middleware/session/token/jwt/jwt_test.go
@@ -15,7 +15,7 @@ func TestDecoing(t *testing.T) {
 	data := map[string]any{
 		"one": float64(1.0),
 		"two": "two",
-		"exp": time.Now().Add(time.Hour).Unix(),
+		"exp": float64(time.Now().Add(time.Hour).Unix()),
 	}
 	token, err := tokens.Encode(data)
 	if err != nil {
